tests: introduce pytest helpers
